Add dramsim2_ini files
dramsim2 and the emulator always look for these specific files in
the current working directory. Whilst it would be better to fix
the emulator so that it doesn't do this, it would require changes
across multiple git repos and be a bit fiddly / time consuming, so
I'm taking the easier route.